requirements.txt
torch
transformers
bottle
grpcio
grpcio-tools
Cython